refactor - MainActivity 수정.
